update with new sdk for new 4.4.0 features

+++ src/dctl-contract-rm.js
@@ -6,16 +6,25 @@ program
     smartContractIds: 'The ID of the contract(s) you want to delete'
   })
   .arguments('<smartContractIds...>')
+  .option('-t, --transactionType [transactionType]', '(mutually exclusive with smartContractId) Transaction type of the contract to delete')
   .option('-v, --verbose', '(optional) Enable STDOUT logger in your Dragonchain SDK')
   .option('-i, --dragonchainId [dragonchainID]', '(optional) Override the default dragonchain ID for this command')
   .parse(process.argv);
 
 util.wrapper(program, async client => {
+  const { transactionType } = program;
+  if (program.args.length !== 0 && transactionType) throw new Error("Please provide only 1 of 'smartContractIds' or 'transactionType'");
   const results = [];
-  await Promise.all(
-    program.args.map(async smartContractId => {
-      results.push(await client.deleteSmartContract({ smartContractId }));
-    })
-  );
+  if (program.args.length > 0) {
+    await Promise.all(
+      program.args.map(async smartContractId => {
+        results.push(await client.deleteSmartContract({ smartContractId }));
+      })
+    );
+  } else if (transactionType) {
+    results.push(await client.deleteSmartContract({ transactionType }));
+  } else {
+    throw new Error("At least one of 'smartContractIds' or 'transactionType' must be provided");
+  }
   console.log(JSON.stringify(results, null, 2));
 });
